+++ driver/src/dma/xaie_dma_aie.h
@@ -32,13 +32,13 @@ void _XAie_ShimDmaInit(XAie_DmaDesc *Desc);
 AieRC _XAie_DmaSetLock(XAie_DmaDesc *DmaDesc, XAie_Lock Acq, XAie_Lock Rel,
 		u8 AcqEn, u8 RelEn);
 AieRC _XAie_ShimDmaWriteBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAie_ShimDmaReadBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAie_TileDmaWriteBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAie_TileDmaReadBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAie_DmaSetInterleaveEnable(XAie_DmaDesc *DmaDesc, u8 DoubleBuff,
 		u8 IntrleaveCount, u16 IntrleaveCurr);
 AieRC _XAie_DmaSetMultiDim(XAie_DmaDesc *DmaDesc, XAie_DmaTensor *Tensor);
@@ -53,14 +53,14 @@ AieRC _XAie_DmaGetChannelStatus(XAie_DevInst *DevInst, XAie_LocType Loc,
 		u32 *Status);
 AieRC _XAie_DmaCheckBdChValidity(u8 BdNum, u8 ChNum);
 AieRC _XAie_DmaUpdateBdLen(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u32 Len, u8 BdNum);
+		XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAie_ShimDmaUpdateBdLen(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u32 Len, u8 BdNum);
+		XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAie_DmaUpdateBdAddr(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u64 Addr, u8 BdNum);
+		XAie_LocType Loc, u64 Addr, u16 BdNum);
 AieRC _XAie_ShimDmaUpdateBdAddr(XAie_DevInst *DevInst,
 		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u64 Addr,
-		u8 BdNum);
+		u16 BdNum);
 AieRC _XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
 		u8 IterCurr);
 AieRC _XAie_AxiBurstLenCheck(u8 BurstLen);
