+++ driver/src/dma/xaie_dma_aieml.h
@@ -34,17 +34,17 @@ void _XAieMl_MemTileDmaInit(XAie_DmaDesc *Desc);
 AieRC _XAieMl_DmaSetLock(XAie_DmaDesc *DmaDesc, XAie_Lock Acq, XAie_Lock Rel,
 		u8 AcqEn, u8 RelEn);
 AieRC _XAieMl_MemTileDmaWriteBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_MemTileDmaReadBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_TileDmaWriteBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_TileDmaReadBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_ShimDmaWriteBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_ShimDmaReadBd(XAie_DevInst *DevInst , XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 AieRC _XAieMl_DmaSetMultiDim(XAie_DmaDesc *DmaDesc, XAie_DmaTensor *Tensor);
 AieRC _XAieMl_DmaGetPendingBdCount(XAie_DevInst *DevInst, XAie_LocType Loc,
 		const XAie_DmaMod *DmaMod, u8 ChNum, XAie_DmaDirection Dir,
@@ -61,14 +61,14 @@ AieRC _XAieMl_DmaGetChannelStatus(XAie_DevInst *DevInst, XAie_LocType Loc,
 AieRC _XAieMl_DmaCheckBdChValidity(u8 BdNum, u8 ChNum);
 AieRC _XAieMl_MemTileDmaCheckBdChValidity(u8 BdNum, u8 ChNum);
 AieRC _XAieMl_DmaUpdateBdLen(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u32 Len, u8 BdNum);
+		XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAieMl_ShimDmaUpdateBdLen(XAie_DevInst *DevInst,
-		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u32 Len, u8 BdNum);
+		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAieMl_DmaUpdateBdAddr(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u64 Addr, u8 BdNum);
+		XAie_LocType Loc, u64 Addr, u16 BdNum);
 AieRC _XAieMl_ShimDmaUpdateBdAddr(XAie_DevInst *DevInst,
 		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u64 Addr,
-		u8 BdNum);
+		u16 BdNum);
 AieRC _XAieMl_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
 		u8 IterCurr);
